Install eclim on Mac

+++ install.sh
@@ -34,9 +34,11 @@ else
 fi
 
 # .vim style
+eclipseHome=$HOME/.vim/eclipse
 if [[ `uname` == 'Darwin' ]]; then
     cp -R vim/ ~/.vim
-else 
+    eclipseHome=$HOME/.vim/Eclipse.app/Contents/Eclipse
+else
     cp -rT vim ~/.vim
 fi
 
@@ -80,13 +82,30 @@ fi
 # http://www.eclipse.org/downloads/packages/eclipse-ide-java-ee-developers/mars2
 if [ "x$JAVA" != "x" ] || [ "x$UPDATE_JAVA_CONFIG" != "x" ]; then
     echo "If $HOME/.vim/eclipse is exists, skipping installing eclipse"
-    if [ ! -d $HOME/.vim/eclipse ]; then
-        # Add hostname into /etc/hosts. e.g. 127.0.0.1 xxxHost
-        #http://www.eclipse.org/downloads/packages/eclipse-ide-java-ee-developers/mars2
-        cd ~/ && wget http://mirror.downloadvn.com/eclipse//technology/epp/downloads/release/mars/2/eclipse-jee-mars-2-linux-gtk-x86_64.tar.gz
-        if [ -d $HOME/.vim/eclipse ]; then rm -rf ~/.vim/eclipse; fi
-        cd ~/ && tar -zxvf eclipse-jee-mars-2-linux-gtk-x86_64.tar.gz
-        mv ~/eclipse ~/.vim/
+    if [[ `uname` == 'Darwin' ]]; then
+        wgetPath=`which wget`
+        if [ "x$wgetPath" == "x" ];then
+            echo "You have to install wget"
+            exit 1
+        fi
+        if [ ! -d $HOME/.vim/Eclipse.app ]; then
+            # Add hostname into /etc/hosts. e.g. 127.0.0.1 xxxHost
+            #http://www.eclipse.org/downloads/packages/eclipse-ide-java-ee-developers/mars2
+            cd ~/ && wget http://mirror.rise.ph/eclipse/technology/epp/downloads/release/mars/2/eclipse-jee-mars-2-macosx-cocoa-x86_64.tar.gz
+            if [ -d $HOME/.vim/eclipse ]; then rm -rf ~/.vim/eclipse; fi
+            cd ~/ && tar -zxvf eclipse-jee-mars-2-macosx-cocoa-x86_64.tar.gz
+            mv ~/Eclipse.app ~/.vim/
+        fi
+
+    else
+        if [ ! -d $HOME/.vim/eclipse ]; then
+            # Add hostname into /etc/hosts. e.g. 127.0.0.1 xxxHost
+            #http://www.eclipse.org/downloads/packages/eclipse-ide-java-ee-developers/mars2
+            cd ~/ && wget http://mirror.downloadvn.com/eclipse//technology/epp/downloads/release/mars/2/eclipse-jee-mars-2-linux-gtk-x86_64.tar.gz
+            if [ -d $HOME/.vim/eclipse ]; then rm -rf ~/.vim/eclipse; fi
+            cd ~/ && tar -zxvf eclipse-jee-mars-2-linux-gtk-x86_64.tar.gz
+            mv ~/eclipse ~/.vim/
+        fi
     fi
 
     if [ ! -d "$HOME/workspace/.metadata/.plugins/org.eclipse.core.runtime/.settings/" ];then
@@ -104,25 +123,39 @@ fi
 
 cd $pwd
 if [ "x$JAVA" != "x" ]; then
+    #install Xvfb
+    if [[ `uname` != 'Darwin' ]]; then
+        sudo yum install xorg-x11-server-Xvfb -y
+    fi
+
     if [ -f "/usr/share/X11/xkb/symbols/inet" ]; then
         sudo sed 's/\/\/zz//' -i /usr/share/X11/xkb/symbols/inet
         sudo sed 's/key <FK20>.*XF86AudioMicMute/\/\/zzkey <FK20> {[XF86AudioMicMute/' -i /usr/share/X11/xkb/symbols/inet
     fi
     ps aux |grep -i Xvfb |grep -v grep | awk '{printf "kill -9 %s\n",$2}' | sudo sh
     export DISPLAY=:1
-    echo "\n\n=== Start Xvfb ===\n\n"
-    sudo Xvfb :1 -screen 0 1024x768x24 &
 
+    if [[ `uname` != 'Darwin' ]]; then
+        echo "\n\n=== Start Xvfb ===\n\n"
+        sudo Xvfb :1 -screen 0 1024x768x24 &
+    fi
     echo "\n\n=== Start eclipse ===\n\n"
-    DISPLAY=:1 ~/.vim/eclipse/eclipse -nosplash -consolelog -debug -application org.eclipse.equinox.p2.director   -repository http://download.eclipse.org/releases/juno      -installIU org.eclipse.wst.web_ui.feature.feature.group &
+    if [ `uname` == 'Darwin' ]; then
+        open ~/.vim/Eclipse.app
+        echo "\n\n=== Install eclim ===\n\n"
+        java -Dvim.files=$HOME/.vim  -Declipse.home=$eclipseHome  -jar ./vim/javaPlugin/eclim_2.6.0.jar install
+
+    else
+        DISPLAY=:1 ~/.vim/eclipse/eclipse -nosplash -consolelog -debug -application org.eclipse.equinox.p2.director   -repository http://download.eclipse.org/releases/juno      -installIU org.eclipse.wst.web_ui.feature.feature.group &
+        echo "\n\n=== Install eclim ===\n\n"
+        java -Dvim.files=$HOME/.vim  -Declipse.home=$HOME/.vim/eclipse/  -jar ./vim/javaPlugin/eclim_2.6.0.jar install
+
+    fi
     sleep 25
 
 
-    #echo "\n\n=== Install eclim ===\n\n"
-    java -Dvim.files=$HOME/.vim  -Declipse.home=$HOME/.vim/eclipse/  -jar ./vim/javaPlugin/eclim_2.6.0.jar install
-
     #echo "\n\n=== Start eclimd ===\n\n"
-    DISPLAY=:1 ~/.vim/eclipse/eclimd -b
+    DISPLAY=:1 $eclipseHome/eclimd -b
     sleep 20
